Separate project into client and server
