+++ src/client/components/champions/SelectedChampionCard.tsx
@@ -1,10 +1,10 @@
 import { useMantineTheme, MantineStyleProp, Box } from '@mantine/core';
 import { useState } from 'react';
 import { IoCloseOutline } from 'react-icons/io5';
-import { useSelectedTeam } from '../contexts/SelectedTeamContext';
+import { useSelectedTeam } from '../../contexts/SelectedTeamContext';
 import TraitHex from '../traits/TraitHex';
-import { Trait } from '@/types/gameplay/trait';
-import { Champion } from '@/types/gameplay/champion';
+import { Trait } from '@/client/types/gameplay/trait';
+import { Champion } from '@/client/types/gameplay/champion';
 
 interface SelectedChampionCardProps {
   champion: Champion;
