+++ src/client/types/gameplay/trait.ts
